+++ src/router/worryAloneRouter.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { worryAloneController } from "../controller";
-import { auth } from '../middlwares';
+import { auth, validation } from '../middlwares';
 import { body } from "express-validator";
 
 const router: Router = Router();
@@ -8,11 +8,12 @@ const router: Router = Router();
 
 //[POST] worry/alone
 router.post("/",
-[body("title").notEmpty().withMessage("제목이 비었습니다."),
-body("content").notEmpty().withMessage("내용이 비었습니다."), 
-body("categoryId").notEmpty().withMessage("카테고리 아이디가 비었습니다."),
-body("options").notEmpty().withMessage("선택지가 비었습니다.")]
+[body("title").notEmpty(),
+body("content").notEmpty(), 
+body("categoryId").notEmpty(),
+body("options").notEmpty()]
 , auth
-,worryAloneController.createAloneWorry);
+, validation
+, worryAloneController.postAloneWorry);
 
 export default router;
